inboxy/pyproject.toml
[project]
name = "inboxy"
version = "1.0.0"
description = "AI-powered Gmail client for the command line"
readme = "README.md"
authors = [
    { name = "ishandhanani", email = "ishandhanani@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.108.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.1.0",
    "openai>=1.6.1",
    "instructor>=1.3.3",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "pydantic>=2.5.3",
    "jinja2>=3.1.2",
    "click>=8.1.7",
]

[project.scripts]
inboxy = "inboxy:main"

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "black>=23.7.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["uv_build>=0.8.6,<0.9.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=inboxy",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]
ignore = []

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=6.3.0",
    "pytest-mock>=3.15.0",
]
